requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
django-cors-headers>=4.3.1
pytz>=2024.1
